requirements.txt
rembg
opencv-python
pillow
numpy
onnxruntime
huggingface_hub
requests
transparent-background
